requirements.txt
web3>=5.27.0
eth-keyfile>=0.5.1
numpy>=1.22.3
pysha3>=1.0.2
eth-brownie>=1.18.0,<2.0.0